# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Dennic Project",
        "contact": {},
        "version": "1.7"
    },
    "host": "localhost:9050",
    "paths": {
...
        "/v1/session": {
            "get": {
                "description": "GetUserSessions - Api for get session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Session"
                ],
                "summary": "GetUserSessions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_session_service.ListSessions"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            },
            "delete": {
                "description": "DeleteSessionById - Api for delete session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Session"
                ],
                "summary": "DeleteSessionById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StatusRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model_common.ResponseError": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model_common.StandardErrorModel": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/model_common.ResponseError"
                }
            }
        },
...
        "model_session_service.ListSessions": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model_session_service.SessionRes"
                    }
                }
            }
        },
        "model_session_service.SessionRes": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "fcm_token": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ip_address": {
                    "type": "string"
                },
                "login_at": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "platform_name": {
                    "type": "string"
                },
                "platform_type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "models.StatusRes": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}